+++ mx/definition/quantization/MaximumFractionalQuantization.h
@@ -17,12 +17,12 @@ public:
   static f64
   QuantizerScaleFactor(const std::array<f64, BlockShape::TotalSize()> &vec) {
     f64 largestValue = 0;
-    for (int i = 0; i < BlockShape::TotalSize(); i++) {
+    for (u32 i = 0; i < BlockShape::TotalSize(); i++) {
       const f64 absValue = fabs(vec[i]);
       largestValue = std::max(largestValue, absValue);
     }
 
-    const u32 scaleFactorCandidate = lround(largestValue);
+    const u32 scaleFactorCandidate = (u32)lround(largestValue);
     const u32 scaleFactorInt = 31 - __builtin_clz(scaleFactorCandidate);
     const u32 scaleFactor = 1 << scaleFactorInt;
